+++ src/Elasticsearch.Net/Elasticsearch.Net.csproj
@@ -72,7 +72,7 @@
     <Compile Include="Domain\Response\VoidResponse.cs" />
     <Compile Include="Domain\Response\ElasticsearchServerError.cs" />
     <Compile Include="Exceptions\OneToOneServerException.cs" />
-    <Compile Include="Exceptions\ElasticsearchAuthenticationException.cs" />
+    <Compile Include="Exceptions\ElasticsearchAuthException.cs" />
     <Compile Include="IElasticsearchClient.cs" />
     <Compile Include="Obsolete\IndicesDeleteAlias.cs" />
     <Compile Include="Obsolete\IndicesPutAlias.cs" />
